Created repeated.rs
