+++ crates/typst/src/layout/grid/mod.rs
@@ -1,6 +1,7 @@
 mod cells;
 mod layout;
 mod lines;
+mod repeated;
 mod rowspans;
 
 pub use self::cells::{
