+++ crates/typst/src/layout/grid/rowspans.rs
@@ -1,3 +1,5 @@
+use super::layout::{in_last_with_offset, points, Row, RowPiece};
+use super::repeated::Repeatable;
 use crate::diag::SourceResult;
 use crate::engine::Engine;
 use crate::foundations::Resolve;
@@ -5,8 +7,6 @@ use crate::layout::{
     Abs, Axes, Cell, Frame, GridLayouter, LayoutMultiple, Point, Regions, Size, Sizing,
 };
 use crate::util::MaybeReverseIter;
-
-use super::layout::{in_last_with_offset, points, Repeatable, Row, RowPiece};
 
 /// All information needed to layout a single rowspan.
 pub(super) struct Rowspan {
